[MIGRATO-164] Добавлен режим импорта --safe-delete при котором не удаляются записи, которых нет в XML файлах

+++ lib/tool/console/command/importdatacommand.php
@@ -35,6 +35,7 @@ class ImportDataCommand extends BaseCommand
 		$this->setName('importdata');
 		$this->setHidden(true);
 		$this->setDescription(Loc::getMessage('INTERVOLGA_MIGRATO.IMPORT_DATA_DESCRIPTION'));
+		$this->addOption('safe-delete');
 	}
 
 	public function executeInner()
@@ -52,7 +53,14 @@ class ImportDataCommand extends BaseCommand
 			$this->logNotResolved();
 			$this->analyzeNotImported();
 			$this->deleteMarked();
-			$this->deleteNotImported();
+			if ($this->input->getOption('safe-delete'))
+			{
+				$this->displayNotImported();
+			}
+			else
+			{
+				$this->deleteNotImported();
+			}
 			$this->resolveReferences();
 		}
 		PublicCache::getInstance()->clearTagCache();
@@ -424,6 +432,21 @@ class ImportDataCommand extends BaseCommand
 		foreach ($this->list->getRecordsToDelete() as $dataRecord)
 		{
 			$this->deleteRecordWithLog($dataRecord);
+		}
+	}
+
+	protected function displayNotImported()
+	{
+		$this->logger->startStep(Loc::getMessage('INTERVOLGA_MIGRATO.STEP_SKIP_DELETE_NOT_IMPORTED'));
+		foreach ($this->list->getRecordsToDelete() as $dataRecord)
+		{
+			$this->logger->addDb(
+				array(
+					'RECORD' => $dataRecord,
+					'OPERATION' => Loc::getMessage('INTERVOLGA_MIGRATO.OPERATION_DELETE_SKIPPED'),
+				),
+				Logger::TYPE_OK
+			);
 		}
 	}
